requirements.txt
torch
torchvision
numpy
scikit-learn
opencv-python
pillow
tqdm
matplotlib
django
